+++ opentrons_drivers/common/actions.py
@@ -2,6 +2,7 @@ import time
 from typing import cast
 from typing import Dict, Callable, TypeVar
 from opentrons_drivers.common.custom_types import ActionFn
+from opentrons_drivers.common.methods import LIQUID_METHODS
 from opentrons_drivers.common.custom_types import CoreWell, StaticCtx, JSONType
 from opentrons.protocol_api.labware import Well
 from opentrons.protocol_api.instrument_context import InstrumentContext
@@ -10,137 +11,7 @@ import opentrons_drivers.common.helpers as help
 #---------- Registries of possible functions ----------
 
 ACTION_REGISTRY: dict[str, ActionFn] = {}  # is exported
-LIQUID_METHODS: dict[str, Callable[..., object]] = {}
-
-F = TypeVar("F", bound=Callable[..., object])
-
-def make_registry_decorator(registry: Dict[str, F]) -> Callable[[str], Callable[[F], F]]: # type: ignore[misc]
-    def register(name: str) -> Callable[[F], F]: # type: ignore[misc]
-        def decorator(fn: F) -> F: # type: ignore[misc]
-            registry[name] = fn
-            return fn
-        return decorator
-    return register
-
-register_action = make_registry_decorator(ACTION_REGISTRY)
-register_liquid_method = make_registry_decorator(LIQUID_METHODS)
-
-
-"""
-    All liquid transfer methods must have the same base signature:
-    pipette: InstrumentContext, to: Well, fr: Well, amount: float
-    
-    Anything else can be passed as keyword arguments
-    e.g. rate, iterations, etc. as **kwargs
-
-"""
-@register_liquid_method("basic_liquid_transfer")
-def basic_liquid_transfer(pipette: InstrumentContext, 
-                     to: Well, fr: Well, 
-                     amount: float) -> None:
-    """
-    Basic liquid transfer method.
-
-    Parameters:
-        pipette (InstrumentContext): Pipette to use.
-        to (Well): Target well.
-        fr (Well): Source well.
-        amount (float): Total volume to transfer.
-
-    Returns:
-        None
-    """
-    amts = help.liquid_batching(pipette, amount)
-    for a in amts:
-        pipette.aspirate(a, fr)
-        pipette.air_gap(20)
-        pipette.dispense(location=to.top(z=1))
-        pipette.blow_out(location=to.top(z=1))
-
-
-@register_liquid_method("advanced_liquid_transfer")
-def advanced_liquid_transfer(pipette: InstrumentContext,
-                    to: Well, 
-                    fr: Well, 
-                    amt: float, 
-                    airgap: float, 
-                    touchtip: int, 
-                    blowout: int, 
-                    asprate: float, 
-                    disrate: float
-                    ) -> None:
-        """Perform a liquid transfer between two wells.
-
-        Handles chunking volumes greater than pipette capacity into multiple
-        aspiration/dispense cycles.
-
-        Args:
-            to (Well): Destination well.
-            fr (Well): Source well.
-            amt (float): Volume in µL to transfer.
-            airgap (float): Volume in µL of airgap.
-            touchtip (int): Number of touch-tip cycles.
-            blowout (int): Number of blowout cycles.
-            asprate (float): Aspiration rate multiplier.
-            disrate (float): Dispense rate multiplier.
-
-        Returns:
-            None
-        """
-
-        amts = help.liquid_batching(pipette, amt)
-        for a in amts:
-            can = pipette.max_volume - a
-            initial_ag = min(can * 0.3 * airgap, can)
-            midway_ag = min(can * 0.15 * (1-airgap), can-initial_ag)
-            pipette.aspirate(a, fr.bottom(z=2), rate=asprate) 
-            [pipette.touch_tip(fr) for _ in range(touchtip)] 
-            pipette.air_gap(initial_ag) 
-            pipette.move_to(help.midpoint(fr,to)) 
-            pipette.air_gap(midway_ag, in_place = True) # type: ignore[call-arg]
-            pipette.dispense(a, location=to.top(z=1), rate=disrate) 
-            [pipette.blow_out(location=to.top(z=1)) for _ in range(blowout)]
-            [pipette.touch_tip(to) for _ in range(touchtip)]
-        
-        pipette.move_to(fr.top(z=5))
-
-
-@register_liquid_method("viscous_liquid_transfer")
-def viscous_liquid_transfer(pipette: InstrumentContext, 
-                             to: Well, fr: Well, 
-                             amount: float, 
-                             rate: float) -> None:
-    """
-    Transfer method for viscous liquids.
-
-    Slows down aspiration and dispense speeds, includes touch tips.
-
-    Parameters:
-        pipette (InstrumentContext): Pipette to use.
-        to (Well): Target well.
-        fr (Well): Source well.
-        amount (float): Volume to transfer.
-        rate (float): Aspiration/dispense rate multiplier.
-
-    Returns:
-        None
-    """
-    amts = help.liquid_batching(pipette, amount)
-    for a in amts:
-        pipette.move_to(fr.bottom(z=3))
-        time.sleep(10)
-        pipette.aspirate(a, fr.bottom(z=3), rate=rate)
-        [pipette.touch_tip(radius=0.5, speed=30, v_offset=-50) for _ in range(3)]
-        pipette.move_to(fr.top())
-        time.sleep(10)
-        [pipette.touch_tip(radius=1, speed=30, v_offset=-10) for _ in range(3)]
-        pipette.dispense(a, location=to.top(z=1), rate=rate)
-        pipette.touch_tip(radius=1, speed=400, v_offset=-5)
-        pipette.dispense(a, location=fr.top(z=-1), rate=rate)
-        pipette.blow_out(location=fr.top(z=1))
-
-
-#---------- Complex exportable functions ----------
+register_action = help.make_registry_decorator(ACTION_REGISTRY)
 
 """
     All exportable functions must have the same base signature:
@@ -308,7 +179,7 @@ def transfer_execution(ctx: StaticCtx, arg: dict[str, JSONType]) -> bool:
 @register_action("sampler_action")
 def sampler_action(ctx: StaticCtx, arg: dict[str, JSONType]) -> bool:
     pip = ctx["pipettes"][arg.get("sampler_mount", "left")]
-    state = ctx["agent_state"]
+    state = ctx["system_state"]
 
     mode = arg.get("mode")
     if mode not in {"scan", "wash", "lift"}:
